# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Umbrella Industrial Services ERP API",
        "description": "Documentation API de l'ERP Umbrella",
        "contact": {
            "email": "support@umbrella.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/agent": {
            "get": {
                "tags": [
                    "Agents"
                ],
                "summary": "Liste de tous les agents",
                "operationId": "39cb7f7e8bbb0ad49b14dbd4fd6a7bd3",
                "responses": {
                    "200": {
                        "description": "Liste des agents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Agent"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Agents"
                ],
                "summary": "Créer un nouvel agent",
                "operationId": "75886c27781ce381f38b0c570a229b80",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Agent créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Agent"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/agent/{id}": {
            "get": {
                "tags": [
                    "Agents"
                ],
                "summary": "Afficher un agent spécifique",
                "operationId": "fed00983b2f90c7702bf0391bdd40e03",
                "parameters": [
                    {
                        "name": "agent",
                        "in": "path",
                        "description": "ID de l'agent",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'agent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Agent"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Agents"
                ],
                "summary": "Mettre à jour un agent",
                "operationId": "41a512ce00a1f60aaf2387249ee2e183",
                "parameters": [
                    {
                        "name": "agent",
                        "in": "path",
                        "description": "ID de l'agent",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Agent mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Agent"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Agents"
                ],
                "summary": "Supprimer un agent",
                "operationId": "1cc42dc7b60a6bfff9d8aa613dc16252",
                "parameters": [
                    {
                        "name": "agent",
                        "in": "path",
                        "description": "ID de l'agent",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agent supprimé avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion utilisateur",
                "operationId": "ea134134deae11eb974c719450ef8a02",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "matricule",
                                    "password"
                                ],
                                "properties": {
                                    "matricule": {
                                        "type": "string",
                                        "example": "UIS00042"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    },
                    "403": {
                        "description": "Compte verrouillé"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Déconnexion utilisateur",
                "operationId": "5ebea01ab5f21df71d38f3505718a1d7",
                "responses": {
                    "200": {
                        "description": "Déconnecté avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Agent": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "matricule": {
                        "type": "string",
                        "example": "AGT12345"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "location": {
                        "type": "string",
                        "example": "Yaoundé"
                    },
                    "phone": {
                        "type": "string",
                        "example": "699123456"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "agent"
                    },
                    "site": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Site A"
                            },
                            "location": {
                                "type": "string",
                                "example": "Douala"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "AgentRequest": {
                "required": [
                    "first_name",
                    "last_name",
                    "location",
                    "email",
                    "site_id",
                    "password"
                ],
                "properties": {
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "location": {
                        "type": "string",
                        "example": "Yaoundé"
                    },
                    "phone": {
                        "type": "string",
                        "example": "699123456"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "site_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "password": {
                        "type": "string",
                        "example": "secret123"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "secret123"
                    }
                },
                "type": "object"
            }
        }
    },
...
}